Adicionando comentários e padrões no código.
