+++ BoletoCloudApi/src/BoletoCloudApi.Api/Controllers/BoletosController.cs
@@ -1,20 +1,28 @@
-﻿using AutoMapper;
-using BoletoCloudApi.Api.ViewModels;
-using BoletoCloudApi.Business.Interfaces;
-using BoletoCloudApi.Business.Models;
-using BoletoCloudApi.Business.Services;
-using Microsoft.AspNetCore.Mvc;
-using System.Net;
-
-namespace BoletoCloudApi.Api.Controllers
+﻿namespace BoletoCloudApi.Api.Controllers
 {
+    using System.Net;
+    using AutoMapper;
+    using BoletoCloudApi.Api.ViewModels;
+    using BoletoCloudApi.Business.Interfaces;
+    using BoletoCloudApi.Business.Models;
+    using BoletoCloudApi.Business.Services;
+    using Microsoft.AspNetCore.Mvc;
+
+    /// <summary>
+    /// Controller responsável por gerenciar operações relacionadas a boletos.
+    /// </summary>
     [Route("api/boletos")]
     public class BoletosController : MainController
     {
         private readonly IMapper _mapper;
         private readonly IBoletoRepository _boletosRepository;
         private readonly IBoletoService _boletoService;
-        
+
+        /// <summary>
+        /// Obtém um boleto pelo identificador.
+        /// </summary>
+        /// <param name="id">Identificador do boleto.</param>
+        /// <returns>O <see cref="BoletoViewModel"/> correspondente ou NotFound.</returns>
         public BoletosController(IMapper mapper,
                                  IBoletoRepository boletosRepository,
                                  IBoletoService boletoService,
@@ -25,12 +33,21 @@ namespace BoletoCloudApi.Api.Controllers
             _boletoService = boletoService;
         }
 
+        /// <summary>
+        /// Obtém todos os boletos cadastrados.
+        /// </summary>
+        /// <returns>Lista de <see cref="BoletoViewModel"/>.</returns>
         [HttpGet]
         public async Task<IEnumerable<BoletoViewModel>> ObterTodos()
         {
             return _mapper.Map<IEnumerable<BoletoViewModel>>(await _boletosRepository.ObterTodos());
         }
 
+        /// <summary>
+        /// Obtém um boleto pelo identificador.
+        /// </summary>
+        /// <param name="id">Identificador do boleto.</param>
+        /// <returns>O <see cref="BoletoViewModel"/> correspondente ou NotFound.</returns>
         [HttpGet("{id:guid}")]
         public async Task<ActionResult<BoletoViewModel>> ObterPorId(Guid id)
         {
@@ -41,6 +58,14 @@ namespace BoletoCloudApi.Api.Controllers
             return boleto;
         }
 
+        /// <summary>
+        /// Busca boletos por número, CPF, e intervalo de datas de emissão.
+        /// </summary>
+        /// <param name="numero">Número do boleto.</param>
+        /// <param name="cpf">CPF do pagador.</param>
+        /// <param name="DataEmissaoInicio">Data inicial de emissão.</param>
+        /// <param name="DataEmissaoFim">Data final de emissão.</param>
+        /// <returns>Lista de <see cref="BoletoViewModel"/> encontrados ou NotFound.</returns>
         [HttpGet("buscar")]
         public async Task<ActionResult<IEnumerable<BoletoViewModel>>> Buscar(
             [FromQuery]string? numero,
@@ -62,6 +87,11 @@ namespace BoletoCloudApi.Api.Controllers
             return Ok(viwModels) ;
         }
 
+        /// <summary>
+        /// Obtém o PDF do boleto pelo identificador.
+        /// </summary>
+        /// <param name="id">Identificador do boleto.</param>
+        /// <returns>Arquivo PDF do boleto ou NotFound.</returns>
         [HttpGet("{id:guid}/pdf")]
         public async Task<ActionResult<byte[]>> ObterBoletoPDFAsync(Guid id)
         {
@@ -79,10 +109,15 @@ namespace BoletoCloudApi.Api.Controllers
             });
         }
 
+        /// <summary>
+        /// Gera um boleto e retorna o PDF para download.
+        /// </summary>
+        /// <param name="boletoViewModel">Dados do boleto a ser gerado.</param>
+        /// <returns>Arquivo PDF do boleto gerado.</returns>
         [HttpPost]
         public async Task<ActionResult<byte[]>> GerarBoletoDownload([FromBody] BoletoViewModel boletoViewModel)
         {
-            if (!ModelState.IsValid) 
+            if (!ModelState.IsValid)
                 return CustomResponse(ModelState);
 
             var boleto = _mapper.Map<Boleto>(boletoViewModel);
