+++ BoletoCloudApi/src/BoletoCloudApi.Api/ViewModels/BoletoViewModel.cs
@@ -1,121 +1,82 @@
-﻿using System.ComponentModel.DataAnnotations;
-
-namespace BoletoCloudApi.Api.ViewModels
+﻿namespace BoletoCloudApi.Api.ViewModels
 {
+    using System.ComponentModel.DataAnnotations;
+
+    /// <summary>
+    /// view model para representar um boleto.
+    /// </summary>
     public class BoletoViewModel
     {
+        /// <summary>
+        /// Identificador único do boleto.
+        /// </summary>
         public Guid Id { get; set; }
 
+        /// <summary>
+        /// Número unico do boleto.
+        /// </summary>
         [Required(ErrorMessage = "O campo {0} é obrigatório")]
         public string Numero { get; set; }
 
+        /// <summary>
+        /// Identificador da conta bancária associada ao boleto.
+        /// </summary>
         [Required(ErrorMessage = "O campo {0} é obrigatório")]
         public string Documento { get; set; }
-        
+
+        /// <summary>
+        /// Numero sequencial associada ao boleto.
+        /// </summary>
         [Required(ErrorMessage = "O campo {0} é obrigatório")]
         public int Sequencial { get; set; }
-        
+
+        /// <summary>
+        /// Valor associada ao boleto.
+        /// </summary>
         [Required(ErrorMessage = "O campo {0} é obrigatório")]
         public decimal Valor { get; set; }
-        
+
+        /// <summary>
+        /// Data de vencimento do boleto.
+        /// </summary>
         [Required(ErrorMessage = "O campo {0} é obrigatório")]
         public DateTime Vencimento { get; set; }
-        
+
+        /// <summary>
+        /// Data de emissão do boleto.
+        /// </summary>
         [Required(ErrorMessage = "O campo {0} é obrigatório")]
         public DateTime Emissao { get; set; }
-        
+
+        /// <summary>
+        /// Titulo do boleto.
+        /// </summary>
         [Required(ErrorMessage = "O campo {0} é obrigatório")]
         public string Titulo { get; set; }
-        
+
+        /// <summary>
+        /// Token único gerado do boleto.
+        /// </summary>
         public string? Token { get; set; }
+
+        /// <summary>
+        /// Data de criação do boleto.
+        /// </summary>
         public DateTime CriadoEm { get; set; }
 
+        /// <summary>
+        /// Conta do boleto.
+        /// </summary>
         public ContaBancariaViewModel? Conta { get; set; }
+
+        /// <summary>
+        /// Beneficiario do boleto.
+        /// </summary>
         public BeneficiarioViewModel? Beneficiario { get; set; }
+
+        /// <summary>
+        /// Pagador do boleto.
+        /// </summary>
         public PagadorViewModel? Pagador { get; set; }
-    }
-
-    public class ContaBancariaViewModel
-    {
-        public Guid Id { get; set; }
-
-        [Required(ErrorMessage = "O campo {0} é obrigatório")]
-        public string Numero { get; set; }
-        
-        [Required(ErrorMessage = "O campo {0} é obrigatório")]
-        public int Sequencial { get; set; }
-        
-        [Required(ErrorMessage = "O campo {0} é obrigatório")]
-        public string Banco { get; set; }
-
-        [Required(ErrorMessage = "O campo {0} é obrigatório")]
-        public string Agencia { get; set; }
-
-        [Required(ErrorMessage = "O campo {0} é obrigatório")]
-        public string Carteira { get; set; }
-        public Guid BoletoId { get; set; }
-    }
-
-    public class BeneficiarioViewModel
-    {
-        public Guid Id { get; set; }
-
-        [Required(ErrorMessage = "O campo {0} é obrigatório")]
-        public string Nome { get; set; }
-
-        [Required(ErrorMessage = "O campo {0} é obrigatório")]
-        [StringLength(14, ErrorMessage = "O campo {0} precisa ter entre {2} e {1} caracteres", MinimumLength = 11)]
-        public string Cprf { get; set; }
-
-        [Required(ErrorMessage = "O campo {0} é obrigatório")]
-        public string Cep { get; set; }
-
-        [Required(ErrorMessage = "O campo {0} é obrigatório")]
-        public string Uf { get; set; }
-
-        [Required(ErrorMessage = "O campo {0} é obrigatório")]
-        public string Localidade { get; set; }
-
-        [Required(ErrorMessage = "O campo {0} é obrigatório")]
-        public string Bairro { get; set; }
-
-        [Required(ErrorMessage = "O campo {0} é obrigatório")]
-        public string Logradouro { get; set; }
-
-        [Required(ErrorMessage = "O campo {0} é obrigatório")]
-        public string Numero { get; set; }
-        public Guid BoletoId { get; set; }
-    }
-
-    public class PagadorViewModel
-    {
-        public Guid Id { get; set; }
-
-        [Required(ErrorMessage = "O campo {0} é obrigatório")]
-        public string Nome { get; set; }
-
-        [Required(ErrorMessage = "O campo {0} é obrigatório")]
-        [StringLength(14, ErrorMessage = "O campo {0} precisa ter entre {2} e {1} caracteres", MinimumLength = 11)]
-        public string CpfCnpj { get; set; }
-
-        [Required(ErrorMessage = "O campo {0} é obrigatório")]
-        public string Cep { get; set; }
-
-        [Required(ErrorMessage = "O campo {0} é obrigatório")]
-        public string Uf { get; set; }
-
-        [Required(ErrorMessage = "O campo {0} é obrigatório")]
-        public string Localidade { get; set; }
-
-        [Required(ErrorMessage = "O campo {0} é obrigatório")]
-        public string Bairro { get; set; }
-
-        [Required(ErrorMessage = "O campo {0} é obrigatório")]
-        public string Logradouro { get; set; }
-
-        [Required(ErrorMessage = "O campo {0} é obrigatório")]
-        public string Numero { get; set; }
-        public Guid BoletoId { get; set; }
-
     }
 }
