+++ BoletoCloudApi/src/BoletoCloudApi.Business/Interfaces/IBoletoRepository.cs
@@ -1,9 +1,17 @@
-﻿using BoletoCloudApi.Business.Models;
-
-namespace BoletoCloudApi.Business.Interfaces
+﻿namespace BoletoCloudApi.Business.Interfaces
 {
+    using BoletoCloudApi.Business.Models;
+
+    /// <summary>
+    /// Interface para operações de persistência e consulta de boletos.
+    /// </summary>
     public interface IBoletoRepository : IRepository<Boleto>
     {
-        Task<Boleto> ObterBoletoPorIdAsync(Guid BoletoId);
+        /// <summary>
+        /// Obtém um boleto pelo identificador informado.
+        /// </summary>
+        /// <param name="boletoId">Identificador único do boleto.</param>
+        /// <returns>O <see cref="Boleto"/> correspondente ao identificador ou <c>null</c> se não encontrado.</returns>
+        Task<Boleto> ObterBoletoPorIdAsync(Guid boletoId);
     }
 }
