+++ BoletoCloudApi/src/BoletoCloudApi.Data/Context/MeuDbContext.cs
@@ -1,21 +1,51 @@
-﻿using BoletoCloudApi.Business.Models;
-using Microsoft.EntityFrameworkCore;
-
-namespace BoletoCloudApi.Data.Context
+﻿namespace BoletoCloudApi.Data.Context
 {
+    using BoletoCloudApi.Business.Models;
+    using Microsoft.EntityFrameworkCore;
+
+    /// <summary>
+    /// Contexto do Entity Framework para acesso ao banco de dados da aplicação.
+    /// Gerencia os conjuntos de entidades e configurações de mapeamento, além de otimizar o rastreamento de alterações.
+    /// </summary>
     public class MeuDbContext : DbContext
     {
+        /// <summary>
+        /// Inicializa uma nova instância do contexto <see cref="MeuDbContext"/>.
+        /// Configura o rastreamento de consultas para não rastrear entidades e desabilita a detecção automática de alterações.
+        /// </summary>
+        /// <param name="options">Opções de configuração do contexto.</param>
         public MeuDbContext(DbContextOptions<MeuDbContext> options) : base(options)
         {
             ChangeTracker.QueryTrackingBehavior = QueryTrackingBehavior.NoTracking;
             ChangeTracker.AutoDetectChangesEnabled = false;
         }
 
+        /// <summary>
+        /// Conjunto de entidades <see cref="Boleto"/>.
+        /// </summary>
         public DbSet<Boleto> Boletos { get; set; }
+
+        /// <summary>
+        /// Conjunto de entidades <see cref="ContaBancaria"/>.
+        /// </summary>
         public DbSet<ContaBancaria> ContasBancarias { get; set; }
+
+        /// <summary>
+        /// Conjunto de entidades <see cref="Beneficiario"/>.
+        /// </summary>
         public DbSet<Beneficiario> Beneficiarios { get; set; }
+
+        /// <summary>
+        /// Conjunto de entidades <see cref="Pagador"/>.
+        /// </summary>
         public DbSet<Pagador> Pagadores { get; set; }
 
+        /// <summary>
+        /// Configura o mapeamento das entidades e propriedades do modelo.
+        /// Define o tipo das colunas string como varchar(100), aplica configurações do assembly
+        /// e ajusta o comportamento de deleção para relacionamentos.
+        /// </summary>
+        /// <param name="modelBuilder">Construtor do modelo de entidades.</param>
         protected override void OnModelCreating(ModelBuilder modelBuilder)
         {
             foreach (var property in modelBuilder.Model.GetEntityTypes()
@@ -29,24 +59,5 @@ namespace BoletoCloudApi.Data.Context
 
             base.OnModelCreating(modelBuilder);
         }
-
-        //public override Task<int> SaveChangesAsync(CancellationToken cancellationToken = default)
-        //{
-        //    foreach (var entry in ChangeTracker.Entries().Where(entry => entry.Entity.GetType().GetProperty("DataCadastro") != null))
-        //    {
-        //        if (entry.State == EntityState.Added)
-        //        {
-        //            entry.Property("DataCadastro").CurrentValue = DateTime.Now;
-        //        }
-
-        //        if (entry.State == EntityState.Modified)
-        //        {
-        //            entry.Property("DataCadastro").IsModified = false;
-        //        }
-        //    }
-
-        //    return base.SaveChangesAsync(cancellationToken);
-        //}
-
     }
 }
