+++ BoletoCloudApi/src/BoletoCloudApi.Data/Mappings/ContaBancariaMapping.cs
@@ -1,16 +1,20 @@
-﻿using BoletoCloudApi.Business.Models;
-using Microsoft.EntityFrameworkCore;
-using Microsoft.EntityFrameworkCore.Metadata.Builders;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace BoletoCloudApi.Data.Mappings
+﻿namespace BoletoCloudApi.Data.Mappings
 {
+    using BoletoCloudApi.Business.Models;
+    using Microsoft.EntityFrameworkCore;
+    using Microsoft.EntityFrameworkCore.Metadata.Builders;
+
+    /// <summary>
+    /// Configuração de mapeamento da entidade <see cref="ContaBancaria"/> para o Entity Framework.
+    /// Define propriedades, tipos de coluna e restrições para persistência no banco de dados.
+    /// </summary>
     public class ContaBancariaMapping : IEntityTypeConfiguration<ContaBancaria>
     {
+        /// <summary>
+        /// Configura o mapeamento da entidade <see cref="ContaBancaria"/>,
+        /// especificando chave primária, tipos de coluna, obrigatoriedade e nome da tabela.
+        /// </summary>
+        /// <param name="builder">Construtor de configuração para a entidade.</param>
         public void Configure(EntityTypeBuilder<ContaBancaria> builder)
         {
             builder.HasKey(x => x.Id);
@@ -27,7 +31,6 @@ namespace BoletoCloudApi.Data.Mappings
             builder.Property(x => x.Carteira)
                 .IsRequired()
                 .HasColumnType("varchar(20)");
-            
             builder.ToTable("ContaBancarias");
         }
     }
